fix: refactor and encapsulate kserve feat in file
